Project Files/requirements.txt
streamlit==1.54.0
google-genai==1.63.0
pillow==12.1.1
